Bug fixes (#58)
* tmp set teambrella service to the main thread

* fix ClaimVotes list vote representation from Double to percents

* fix voting from ClaimVC

* fix number view

* when creating new claim push to that claim

* minor bug fixes

* >

* add transition to application via teambrella qr-code

* hide QR code login option when user has real private key

+++ App/Source/Scenes/Login/LoginBlueVC.swift
@@ -32,10 +32,10 @@ final class LoginBlueVC: UIViewController {
     @IBOutlet var qrCodeButton: UIButton!
     @IBOutlet var gradientView: GradientView!
     @IBOutlet var confetti: SKView!
-
+    
     var isEmitterAdded: Bool = false
     var didTapDemo: Bool = false
-
+    
     var loginWorker: LoginWorker!
     
     lazy var secretRecognizer: UILongPressGestureRecognizer = {
@@ -52,15 +52,19 @@ final class LoginBlueVC: UIViewController {
     
     var isRegisteredFacebookUser: Bool { return KeychainService().value(forKey: .privateKey) != nil }
     
+    var canLogWithQRCode: Bool {
+        return !service.keyStorage.isRealPrivateKeySet || SimpleStorage().bool(forKey: .isRegistering)
+    }
+    
     // MARK: Lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
+        
         if service.joinTeamID != nil {
             performSegue(type: .welcome)
         }
-
+        
         centerLabel.text = "Login.LoginBlueVC.centerLabel".localized
         continueWithFBButton.setTitle("Login.LoginBlueVC.continueWithFBButton".localized, for: .normal)
         continueWithVKButton.setTitle("Login.LoginBlueVC.continueWithVKButton".localized, for: .normal)
@@ -74,22 +78,24 @@ final class LoginBlueVC: UIViewController {
         
         centerLabel.isUserInteractionEnabled = true
         centerLabel.addGestureRecognizer(secretRecognizer)
-
-        qrCodeButton.addGestureRecognizer(clearAllRecognizer)
+        
+        nextButton.addGestureRecognizer(clearAllRecognizer)
         //        continueWithFBButton.addGestureRecognizer(clearAllRecognizer)
         animateCenterLabel()
-
+        
         NotificationCenter.default.addObserver(self,
                                                selector: #selector(dynamicLinkReceived),
                                                name: .dynamicLinkReceived,
                                                object: nil)
     }
-
+    
     override func viewWillAppear(_ animated: Bool) {
         
         super.viewWillAppear(animated)
         centerLabel.alpha = 0
         gradientView.alpha = 0
+        
+        qrCodeButton.isHidden = !canLogWithQRCode
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -114,7 +120,7 @@ final class LoginBlueVC: UIViewController {
     @IBAction func tapContinueWithVKButton() {
         register(type: .vk)
     }
-
+    
     @IBAction func tapNextButton() {
         if service.keyStorage.isRealPrivateKeySet && SimpleStorage().bool(forKey: .isRegistering) == false {
             service.keyStorage.setToRealUser()
@@ -130,7 +136,7 @@ final class LoginBlueVC: UIViewController {
             logIn()
             return
         }
-
+        
         HUD.show(.progress)
         loginWorker.loginAndRegister(type: type, in: self) { [weak self] token, error in
             if token != nil && error == nil {
@@ -145,20 +151,20 @@ final class LoginBlueVC: UIViewController {
         service.keyStorage.setToDemoUser()
         performSegue(type: .unwindToInitial, sender: self)
     }
-
+    
     @IBAction func tapQRCode(_ sender: UIButton) {
         Statistics.log(event: .tapQRCodeLogin)
         service.router.showCodeCapture(in: self, delegate: self, type: .privateKey)
     }
-
-     @IBAction func unwindToLogin(segue: UIStoryboardSegue) {
+    
+    @IBAction func unwindToLogin(segue: UIStoryboardSegue) {
         if let vc = segue.source as? ApplicationFlowVC, let error = vc.error {
             DispatchQueue.main.asyncAfter(deadline: .now() + 0.25) {
                 self.handleFailure(error: error)
             }
         }
     }
-
+    
     @objc
     private func secretTap(sender: UILongPressGestureRecognizer) {
         let controller = UIAlertController(title: "Secret entrance",
@@ -189,11 +195,19 @@ Are you sure you want to completely remove your private key from this device?
 """,
                                            preferredStyle: .alert)
         
-        controller.addAction(UIAlertAction(title: "Yes", style: .destructive, handler: { action in
+        controller.addAction(UIAlertAction(title: "Yes", style: .destructive, handler: { [weak self] action in
+            guard let self = self else { return }
+            
             service.keyStorage.deleteStoredKeys()
+            let animator = UIViewPropertyAnimator(duration: 0.3, curve: .easeInOut)
+            animator.addAnimations {
+                self.qrCodeButton.isHidden = !self.canLogWithQRCode
+            }
+            animator.startAnimation()
         }))
         controller.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
         
+        Vibrator().lightVibes()
         present(controller, animated: true, completion: nil)
     }
     
@@ -265,18 +279,18 @@ Are you sure you want to completely remove your private key from this device?
         service.keyStorage.saveNewPrivateKey(string: string)
         logIn()
     }
-
+    
     private func logIn() {
         HUD.hide()
         performSegue(type: .unwindToInitial)
     }
-
+    
     @objc
     private func dynamicLinkReceived() {
         navigationController?.popToViewController(self, animated: false)
         performSegue(type: .welcome)
     }
-
+    
     private func newPrivateKeySet(privateKey: String) {
         service.keyStorage.saveNewPrivateKey(string: privateKey)
         service.keyStorage.setToRealUser()
@@ -287,15 +301,34 @@ Are you sure you want to completely remove your private key from this device?
 // MARK: CodeCaptureDelegate
 extension LoginBlueVC: CodeCaptureDelegate {
     func codeCapture(controller: CodeCaptureVC, didCapture: String, type: QRCodeType) {
-        if type == .bitcoinWiF {
+        switch type {
+        case .bitcoinWiF:
             controller.close(cancelled: false)
             self.newPrivateKeySet(privateKey: didCapture)
-        } else {
-            log("Wrong type: \(type)", type: .info)
+        case .surillaLink,
+             .teambrellaLink:
+            log("Is teambrella! \(didCapture)", type: .info)
+            guard canLogWithQRCode else {
+                log("Private key already exists. Universal link transition is cancelled", type: .info)
+                return
+            }
+            guard let url = URL(string: didCapture) else { return }
+            
+            controller.close(cancelled: false)
+            SimpleStorage().store(bool: true, forKey: .isRegistering)
+            let activity = NSUserActivity(activityType: NSUserActivityTypeBrowsingWeb)
+            activity.webpageURL = url
+            let application = UIApplication.shared
+            let delegate = application.delegate
+            _ = delegate?.application?(application, continue: activity, restorationHandler: { item in
+                
+            })
+        default:
+            log("Wrong type: \(type), \(didCapture)", type: .info)
         }
     }
-
+    
     func codeCaptureWillClose(controller: CodeCaptureVC, cancelled: Bool) {
-
+        
     }
 }
